refactor: #12 JavaDoc 작업 및 리팩토링
