+++ src/main/java/com/nhnacademy/trans/service/MqttIngestionService.java
@@ -1,9 +1,10 @@
-package com.nhnacademy.trans;
+package com.nhnacademy.trans.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hivemq.client.mqtt.MqttClient;
 import com.hivemq.client.mqtt.mqtt3.Mqtt3AsyncClient;
 import com.hivemq.client.mqtt.mqtt3.message.publish.Mqtt3Publish;
+import com.nhnacademy.trans.RuleEngine;
 import com.nhnacademy.trans.config.RuleCacheService;
 import com.nhnacademy.trans.domain.Threshold;
 import jakarta.annotation.PostConstruct;
