+++ src/main/resources/application.properties
@@ -16,14 +16,11 @@ eureka.instance.prefer-ip-address=false
 eureka.client.register-with-eureka=true
 server.port=10284
 
-# Redis ??
+# Redis
 spring.data.redis.host=s4.java21.net
 spring.data.redis.port=6379
-# ?? Redis 6+ ACL ???? ??? ?? (?? default)
 spring.data.redis.username=default
-# ?? ??? ????? CI ????? ?????
 spring.data.redis.password=*N2vya7H@muDTwdNMR!
-# ?? ?? DB ?? (optional, ?? 0) ?????????????????????
 spring.data.redis.database=281
