+++ src/test/java/com/nhnacademy/trans/config/InfluxDBConfigTest.java
@@ -1,0 +1,7 @@
+package com.nhnacademy.trans.config;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class InfluxDBConfigTest {
+
+}
